backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.6.4
pydantic[email]==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1
imagehash==4.3.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
faiss-cpu==1.8.0
invisible-watermark==0.2.0
boto3==1.34.82
celery==5.3.6
redis==5.0.3
httpx==0.27.0
sendgrid==6.11.0
python-dotenv==1.0.1
aiofiles==23.2.1
aiohttp==3.9.2
slowapi==0.1.9
loguru==0.7.2
